@@ -1,7 +1,7 @@
 package com.aar.app.wsp.easyadapter;
 
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.ViewGroup;
 
 public abstract class AdapterDelegate<T, VH extends RecyclerView.ViewHolder> {
