@@ -2,8 +2,8 @@ package com.aar.app.wsp.features;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.aar.app.wsp.features.mainmenu.MainMenuActivity;
 
