@@ -1,8 +1,8 @@
 package com.aar.app.wsp.features.settings;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.ActionBar;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.ActionBar;
 import android.view.MenuItem;
 
 import com.aar.app.wsp.R;
