@@ -1,8 +1,8 @@
 package com.aar.app.wsp.data.room;
 
-import android.arch.persistence.room.Dao;
-import android.arch.persistence.room.Insert;
-import android.arch.persistence.room.Query;
+import androidx.room.Dao;
+import androidx.room.Insert;
+import androidx.room.Query;
 
 import com.aar.app.wsp.features.gamethemeselector.GameThemeItem;
 import com.aar.app.wsp.model.GameTheme;
