@@ -1,8 +1,8 @@
 package com.aar.app.wsp.model;
 
-import android.arch.persistence.room.ColumnInfo;
-import android.arch.persistence.room.Entity;
-import android.arch.persistence.room.TypeConverters;
+import androidx.room.ColumnInfo;
+import androidx.room.Entity;
+import androidx.room.TypeConverters;
 
 import com.aar.app.wsp.data.room.AnswerLineConverter;
 
